requirements.txt
Django~=5.0.6
python-telegram-bot~=21.2
cryptography~=42.0.7
